requirements.txt
aiogram>=3.13,<4.0
gspread>=6.1,<7.0
google-auth>=2.35,<3.0
